models/quantile_analysis/requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# ML Models
lightgbm>=3.3.0
xgboost>=1.6.0

# MLflow and Experiment Tracking
mlflow>=2.0.0
optuna>=3.0.0
optuna-integration[mlflow]>=4.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.0.0

# Data Processing
openpyxl>=3.0.0  # For Excel file support
xlrd>=2.0.0      # Additional Excel support

# Optional: Enhanced Features
# databricks-cli>=0.17.0  # For Databricks integration
# boto3>=1.20.0           # For AWS S3 artifact storage
# psycopg2-binary>=2.9.0  # For PostgreSQL backend

# Development and Testing (optional)
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Web Framework (for future dashboard)
flask>=2.0.0
flask-cors>=3.0.0 